// src/models/process-request-image.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Regula Document Reader Web API
 * Regula Document Reader Web API
 *
 * The version of the OpenAPI document: 5.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { ImageData } from './image-data';
import { Light } from './light';

/**
 * 
 * @export
 * @interface ProcessRequestImage
 */
export interface ProcessRequestImage {
    /**
     * 
     * @type {ImageData}
     * @memberof ProcessRequestImage
     */
    ImageData: ImageData;
    /**
     * 
     * @type {Light}
     * @memberof ProcessRequestImage
     */
    LightIndex?: Light;
    /**
     * page/image number
     * @type {number}
     * @memberof ProcessRequestImage
     */
    page_idx?: number;
}